Clean up frontend and some backend code
