+++ views/pages/controlPanel.ejs
@@ -550,6 +550,8 @@
 			}
 		</script>
 		<script src='/js/controlPanelStudents.js'></script>
+		<script src='/js/sockets/cpUpdate.js'></script>
+		<script src='/js/modules/tags.js'></script>
 		<script src='/js/controlPanelSockets.js'></script>
 		<script src='/js/controlPanelPolls.js'></script>
